+++ core/core/langs/asmb_riscv/ir_lines.cpp
@@ -14,7 +14,7 @@ void pepp::tc::IntegerInstruction::insert(std::unique_ptr<AAttribute> attr) {
   } else LinearIR::insert(std::move(attr));
 }
 
-std::optional<u16> pepp::tc::IntegerInstruction::object_size(u16 base_address) const { return 4; }
+std::optional<u64> pepp::tc::IntegerInstruction::object_size(u64 base_address) const { return 4; }
 
 pepp::tc::RTypeIR::RTypeIR(riscv::MnemonicDescriptor desc, u8 rd, u8 rs1, u8 rs2) : IntegerInstruction(desc) {
   this->rd = rd;
